refactor: Move makeSorterFromDefinition to queries
